html restructure
