+++ src/pages/integration-builder/builder/frameworks/html/stepContent/register-app.mdx
